build: rename ui to cnipe.py and move one directory up to make exe files easier
